requirements.txt
asgiref==3.8.1
Babel==2.15.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
currency-codes==23.6.4
defusedxml==0.7.1
Django==4.2.13
django-bootstrap==0.2.4
django-bootstrap-icons==0.8.7
django-bootstrap-v5==1.0.11
django-bootstrap5==24.2
django-countries==7.6.1
django-js-asset==2.2.0
django-money==3.5.2
django-mptt==0.16.0
django-restframework==0.0.1
django-smtp-ssl==1.0
django-widget-tweaks==1.5.0
idna==3.7
iso4217==1.11.20220401
pillow==10.3.0
py-moneyed==3.0
requests==2.31.0
setuptools==70.0.0
soupsieve==2.5
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
